chore: resolve 2 more circular dependencies
