+++ src/components/molecules/MetricsBarChart/MetricsBarChart.tsx
@@ -5,16 +5,9 @@ import debounce from 'lodash.debounce';
 import type {ExecutionMetrics} from '@models/metrics';
 
 import {BarsWrapper, ChartWrapper, MetricsBarChartWrapper} from './MetricsBarChart.styled';
-import Chart from './components/Chart';
+import Chart, {BarChartConfig} from './components/Chart';
 import PAxisLine from './components/PAxisLine';
 import {getAxisPosition, getMaximumValue, getMinimumValue, metricsLogarithmization, secondInMs} from './utils';
-
-export type BarChartConfig = {
-  barWidth: number;
-  barMargin: number;
-  chartHeight: number;
-  chartData: any[];
-};
 
 type MetricsBarChartProps = {
   data?: ExecutionMetrics[];
