+++ src/utils/fetchUtils.ts
@@ -7,6 +7,8 @@ import {isEqual} from 'lodash';
 
 import {getApiEndpoint} from '@services/apiEndpoint';
 
+import {getRtkBaseUrl, getRtkIdToken} from '@utils/rtk';
+
 const prohibitedValues = ['undefined', 'null'];
 
 export const paramsSerializer = (params: object) => {
@@ -28,27 +30,6 @@ export const paramsSerializer = (params: object) => {
       return item !== '';
     })
     .join('&');
-};
-
-type IdTokenResolver = () => Promise<string | null>;
-let resolveIdToken: IdTokenResolver = () => Promise.resolve(null);
-
-type BaseUrlResolver = (routeToRequest: string | undefined) => string;
-let resolveBaseUrl: BaseUrlResolver = () => '';
-export const setRtkIdTokenResolver = (resolver: IdTokenResolver): void => {
-  resolveIdToken = resolver;
-};
-export const setRtkBaseUrlResolver = (resolver: BaseUrlResolver): void => {
-  resolveBaseUrl = resolver;
-};
-
-// TODO: Delete these when not needed.
-//       It's temporarily used to unify ID token and base url strategy.
-export const getRtkBaseUrl = (routeToRequest: string | undefined): string => {
-  return resolveBaseUrl(routeToRequest) || '';
-};
-export const getRtkIdToken = (): Promise<string | null> => {
-  return resolveIdToken();
 };
 
 const rawBaseQuery = (baseUrl: string, idToken?: string | null) =>
@@ -85,8 +66,8 @@ export const dynamicBaseQuery: BaseQueryFn<string | DynamicFetchArgs, unknown, F
     args = {url: args};
   }
 
-  const idToken = await resolveIdToken();
-  const baseUrl = resolveBaseUrl(args.routeToRequest) || '';
+  const idToken = await getRtkIdToken();
+  const baseUrl = getRtkBaseUrl(args.routeToRequest) || '';
 
   return rawBaseQuery(`${apiEndpoint}${baseUrl}`, idToken)(args, api, extraOptions);
 };
